+++ scorex-basics/src/main/scala/scorex/transaction/BlockStorage.scala
@@ -1,5 +1,6 @@
 package scorex.transaction
 
+import org.h2.mvstore.MVStore
 import scorex.block.Block
 import scorex.block.Block.BlockId
 import scorex.crypto.encode.Base58
@@ -12,6 +13,8 @@ import scala.util.{Failure, Success, Try}
   * Storage interface combining both history(blockchain/blocktree) and state
   */
 trait BlockStorage[TX <: Transaction[_]] extends ScorexLogging {
+
+  val db: MVStore
 
   val MaxRollback: Int
 
@@ -26,12 +29,14 @@ trait BlockStorage[TX <: Transaction[_]] extends ScorexLogging {
         state.processBlock(b) match {
           case Failure(e) =>
             log.error("Failed to apply block to state", e)
-            removeAfter(block.referenceField.value)
-            //TODO ???
-            System.exit(1)
+            db.rollback()
           case Success(m) =>
+            db.commit()
         }
       }
+    }.recoverWith { case e =>
+      log.error("Failed to append block:", e)
+      Failure(e)
     }
   }